requirements.txt
msal>=1.28.0
evernote3>=1.25.4
